+++ src/main.js
@@ -3,6 +3,8 @@ import { BootScene } from './scenes/BootScene.js';
 import { MainMenuScene } from './scenes/MainMenuScene.js';
 import { DevOptionsScene } from './scenes/DevOptionsScene.js';
 import { AchievementsScene } from './scenes/AchievementsScene.js';
+import { InGameMenuScene } from './scenes/InGameMenuScene.js';
+import { HelpScene } from './scenes/HelpScene.js';
 import { GameScene } from './scenes/GameScene.js';
 import { UIScene } from './scenes/UIScene.js';
 
@@ -18,7 +20,7 @@ const config = {
     width: window.innerWidth,
     height: window.innerHeight,
   },
-  scene: [BootScene, MainMenuScene, DevOptionsScene, AchievementsScene, GameScene, UIScene],
+  scene: [BootScene, MainMenuScene, DevOptionsScene, AchievementsScene, InGameMenuScene, HelpScene, GameScene, UIScene],
 };
 
 const game = new Phaser.Game(config);